refactor: re-organize file per feature
This brings folder structure closer to the one
recommended in the style guide. This is just the
first step of more moves / renames
