+++ src/app/github/repo-detail/repo-detail.ts
@@ -1,6 +1,6 @@
 import {Component, OnInit} from '@angular/core';
 import {ROUTER_DIRECTIVES, ActivatedRoute, Router} from '@angular/router';
-import {Github} from '../../services/github';
+import {Github} from '../shared/github';
 
 @Component({
   selector: 'repo-detail',
